refactor(ops): rename package and modules in Ops
