+++ src/main/scala/scatan/model/ApplicationState.scala
@@ -3,6 +3,7 @@ package scatan.model
 import scatan.lib.game.{Game, GameRulesDSL, Player}
 import scatan.lib.mvc.Model
 import scatan.model.game.{ScatanActions, ScatanPhases, ScatanRules, ScatanState}
+import scatan.model.game.ScatanActions.*
 
 type ScatanGame = Game[ScatanState, ScatanPhases, ScatanActions]
 type ScatanGameDSL = GameRulesDSL[ScatanState, ScatanPhases, ScatanActions]
