+++ src/main/scala/scatan/model/game/ScatanActions.scala
@@ -6,7 +6,8 @@ import scatan.model.map.Hexagon
 import scatan.model.map.{RoadSpot, StructureSpot}
 import scatan.lib.game.Player
 import scatan.model.components.BuildingType
-import scatan.model.game.state.BuildingCapacity.assignBuilding
+import scatan.model.game.ops.BuildingOps.assignBuilding
+import scatan.model.game.ops.CardOps.assignResourcesFromNumber
 
 enum ScatanActions(effect: ScatanState => ScatanState) extends Action[ScatanState](effect):
   case RollDice(result: Int) extends ScatanActions(ScatanActions.RollEffect(result))
