+++ src/test/scala/scatan/model/game/BasicStateTest.scala
@@ -6,7 +6,7 @@ import scatan.model.game.ScatanState
 import scatan.model.map.Spot
 import scatan.model.map.RoadSpot
 import scatan.model.map.StructureSpot
-import scatan.model.game.state.EmptySpotsManagement.emptySpots
+import scatan.model.game.ops.EmptySpotsOps.emptySpots
 
 abstract class BasicStateTest extends BaseTest:
